Separate Sanbase.Clickhouse.Metric into 2 modules
