+++ lib/sanbase/clickhouse/metric/metric.ex
@@ -11,10 +11,8 @@ defmodule Sanbase.Clickhouse.Metric do
 
   use Ecto.Schema
 
-  import Sanbase.Clickhouse.Metric.Helper,
-    only: [slug_asset_id_map: 0, asset_id_slug_map: 0, metric_name_id_map: 0]
-
-  import Sanbase.DateTimeUtils, only: [round_datetime: 2, str_to_sec: 1]
+  import Sanbase.Clickhouse.Metric.Helper, only: [slug_asset_id_map: 0, asset_id_slug_map: 0]
+  import Sanbase.Clickhouse.Metric.Queries
 
   alias __MODULE__.FileHandler
 
@@ -28,12 +26,10 @@ defmodule Sanbase.Clickhouse.Metric do
   @timeseries_metrics_public_name_list FileHandler.metrics_with_data_type(:timeseries)
   @histogram_metrics_public_name_list FileHandler.metrics_with_data_type(:histogram)
   @access_map FileHandler.access_map()
-  @table_map FileHandler.table_map()
   @min_interval_map FileHandler.min_interval_map()
   @free_metrics FileHandler.metrics_with_access(:free)
   @restricted_metrics FileHandler.metrics_with_access(:restricted)
   @aggregation_map FileHandler.aggregation_map()
-  @name_to_column_map FileHandler.name_to_column_map()
   @human_readable_name_map FileHandler.human_readable_name_map()
   @metrics_data_type_map FileHandler.metrics_data_type_map()
   @metrics_public_name_list (@histogram_metrics_public_name_list ++
@@ -81,8 +77,8 @@ defmodule Sanbase.Clickhouse.Metric do
   end
 
   @impl Sanbase.Metric.Behaviour
-  def histogram_data(metric, slug, from, to, interval) do
-    {query, args} = histogram_data_query(metric, slug, from, to, interval)
+  def histogram_data(metric, slug, from, to, interval, limit) do
+    {query, args} = histogram_data_query(metric, slug, from, to, interval, limit)
 
     ClickhouseRepo.query_transform(query, args, fn [unix, value] ->
       %{
@@ -209,176 +205,5 @@ defmodule Sanbase.Clickhouse.Metric do
           %{slug: Map.get(asset_id_map, asset_id), value: value}
         end)
     end
-  end
-
-  defp aggregation(:last, value_column, dt_column), do: "argMax(#{value_column}, #{dt_column})"
-  defp aggregation(:first, value_column, dt_column), do: "argMin(#{value_column}, #{dt_column})"
-  defp aggregation(aggr, value_column, _dt_column), do: "#{aggr}(#{value_column})"
-
-  defp timeseries_data_query(metric, slug, from, to, interval, aggregation) do
-    query = """
-    SELECT
-      toUnixTimestamp(intDiv(toUInt32(toDateTime(dt)), ?1) * ?1) AS t,
-      #{aggregation(aggregation, "value", "t")}
-    FROM(
-      SELECT
-        dt,
-        argMax(value, computed_at) AS value
-      FROM #{Map.get(@table_map, metric)}
-      PREWHERE
-        dt >= toDateTime(?3) AND
-        dt < toDateTime(?4) AND
-        asset_id = (
-          SELECT argMax(asset_id, computed_at)
-          FROM asset_metadata
-          PREWHERE name = ?5
-        ) AND
-        metric_id = (
-          SELECT
-            argMax(metric_id, computed_at) AS metric_id
-          FROM
-            metric_metadata
-          PREWHERE
-            name = ?2
-        )
-      GROUP BY dt
-    )
-    GROUP BY t
-    ORDER BY t
-    """
-
-    args = [
-      Sanbase.DateTimeUtils.str_to_sec(interval),
-      Map.get(@name_to_column_map, metric),
-      from,
-      to,
-      slug
-    ]
-
-    {query, args}
-  end
-
-  defp histogram_data_query(metric, slug, from, to, interval, limit \\ 100) do
-    query = """
-    SELECT
-      toUnixTimestamp(intDiv(toUInt32(toDateTime(value)), ?5) * ?5) AS t,
-      -sum(measure)
-    FROM #{Map.get(@table_map, metric)} FINAL
-    PREWHERE
-      metric_id = (
-        SELECT argMax(metric_id, computed_at)
-        FROM metric_metadata
-        PREWHERE name = ?1
-      ) AND
-      asset_id = (
-        SELECT argMax(asset_id, computed_at)
-        FROM asset_metadata
-        PREWHERE name = ?2
-      ) AND
-      dt > toDateTime(?3) AND dt < toDateTime(?4) AND dt != value
-    GROUP BY t
-    ORDER BY t DESC
-    LIMIT ?6
-    """
-
-    args = [
-      Map.get(@name_to_column_map, metric),
-      slug,
-      from,
-      to,
-      interval |> str_to_sec(),
-      limit
-    ]
-
-    {query, args}
-  end
-
-  defp aggregated_timeseries_data_query(metric, asset_ids, from, to, aggregation) do
-    query = """
-    SELECT
-      toUInt32(asset_id),
-      #{aggregation(aggregation, "value", "t")}
-    FROM(
-      SELECT
-        dt,
-        asset_id,
-        argMax(value, computed_at) AS value
-      FROM #{Map.get(@table_map, metric)}
-      PREWHERE
-        dt >= toDateTime(?3) AND
-        dt < toDateTime(?4) AND
-        asset_id IN (?1) AND
-        metric_id = ?2
-      GROUP BY dt, asset_id
-    )
-    GROUP BY asset_id
-    """
-
-    {:ok, metric_map} = metric_name_id_map()
-
-    args = [
-      asset_ids,
-      Map.get(metric_map, Map.get(@name_to_column_map, metric)),
-      from,
-      to
-    ]
-
-    {query, args}
-  end
-
-  defp available_slugs_query() do
-    query = """
-    SELECT DISTINCT(name) FROM asset_metadata
-    """
-
-    args = []
-
-    {query, args}
-  end
-
-  defp first_datetime_query(metric, nil) do
-    query = """
-    SELECT
-      toUnixTimestamp(toDateTime(min(dt)))
-    FROM #{@table}
-    PREWHERE
-      metric_id = (
-        SELECT
-          argMax(metric_id, computed_at) AS metric_id
-        FROM
-          metric_metadata
-        PREWHERE
-          name = ?1 ) AND
-      value > 0
-    """
-
-    args = [Map.get(@name_to_column_map, metric)]
-
-    {query, args}
-  end
-
-  defp first_datetime_query(metric, slug) do
-    query = """
-    SELECT
-      toUnixTimestamp(toDateTime(min(dt)))
-    FROM #{@table}
-    PREWHERE
-      asset_id = (
-        SELECT argMax(asset_id, computed_at)
-        FROM asset_metadata
-        PREWHERE name = ?1
-      ) AND metric_id = (
-        SELECT
-          argMax(metric_id, computed_at) AS metric_id
-        FROM
-          metric_metadata
-        PREWHERE
-          name = ?2 ) AND
-      value > 0
-    """
-
-    args = [slug, Map.get(@name_to_column_map, metric)]
-
-    {query, args}
   end
 end
