+++ lib/sanbase/metric/behaviour.ex
@@ -81,5 +81,5 @@ defmodule Sanbase.Metric.Behaviour do
 
   @callback access_map() :: map()
 
-  @optional_callbacks [histogram_data: 3]
+  @optional_callbacks [histogram_data: 6]
 end
